+++ features/pathsfile1.php
@@ -1,0 +1,13 @@
+<?php require_once __DIR__ . '/prepend.php'; ?>
+
+<?php 
+
+  define('PAGE_TITLE', 'Path Test 1');
+
+  renderFeaturesTemplateStart(); 
+ 
+?>
+
+  <p>This page serves no purpose other than for testing URI paths.</p>
+
+<?php renderFeaturesTemplateEnd(); ?>